Bump 3.10.4: modularize skills.ts into 6 focused modules
Split 1878-line skills.ts into a 225-line dispatcher + 6 helpers:
- skills-shipped     foundation readers, constants, resolveDest
- skills-manifest    skill-namespace manifest mutations (prune)
- skills-compile     compile pipeline + collectInstalledSkillNames + rmTreeIfPresent
- skills-user-fs     user-scope path resolvers + readers + host filter
- skills-install     installRepo + installUserSkills + formatters
- skills-status      list/status/validate verbs for both scopes

Apply audit-tail P3-14: collapse two passes over entry.source
(Object.values + Object.keys) to one Object.entries pass in installRepo's
source-missing fallback.

Pure refactor, no behaviour change. Typecheck clean; 46/46 skills tests pass.
