+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "condash",
-  "version": "3.10.3",
+  "version": "3.10.4",
   "description": "Markdown project dashboard for the conception tree (Electron build)",
   "private": true,
   "main": "dist-electron/main/index.js",
